+++ tests/frontendIntegration/drf_sync/mysite/settings.py
@@ -35,15 +35,15 @@ DEBUG = True
 ALLOWED_HOSTS = ["localhost.org", "localhost", "0.0.0.0"]
 
 CORS_ORIGIN_WHITELIST = [
-    "http://localhost.org:8080",
+    "http://localhost:8080",
 ]
 CORS_ALLOW_ALL_ORIGINS = True
 CORS_ALLOW_CREDENTIALS = True
 CORS_ALLOWED_ORIGINS = [
-    "http://localhost.org:8080",
+    "http://localhost:8080",
 ]
 CORS_ALLOWED_ORIGIN_REGEXES = [
-    "http://localhost.org:8080",
+    "http://localhost:8080",
 ]
 
 CORS_ALLOW_METHODS = [
@@ -202,7 +202,7 @@ init(
     app_info=InputAppInfo(
         app_name="SuperTokens Python SDK",
         api_domain="0.0.0.0:" + get_app_port(),
-        website_domain="http://localhost.org:8080",
+        website_domain="http://localhost:8080",
     ),
     framework="django",
     recipe_list=[